requirements.txt
# Flask and extensions
Flask==3.0.0
Flask-SQLAlchemy==3.1.1
Flask-Migrate==4.0.5
python-dotenv==1.0.0

# Database
SQLAlchemy==2.0.25
psycopg2-binary==2.9.9
alembic==1.12.1

# Security
Flask-Login==0.6.3
Flask-WTF==1.2.1
WTForms==3.1.0
email-validator==2.1.0

# Utils
python-dateutil==2.8.2
Pillow==10.1.0

# Testing
pytest==7.4.3
pytest-flask==1.3.0

# Deployment
gunicorn==21.2.0 